backend/requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
python-jose[cryptography]
passlib[bcrypt]
python-multipart
google-auth
pydantic-settings
python-dotenv
redis
pydantic
